Restructured to create login display, login controller and userController
