+++ Hdb/src/UserDisplay.java
@@ -8,7 +8,7 @@ public interface UserDisplay {
         String currentPassword = scanner.nextLine().trim();
         System.out.print("Enter new password: ");
         String newPassword = scanner.nextLine().trim();
-        if (Utils.changePassword(user, currentPassword, newPassword)) {
+        if (UserController.changePassword(user, currentPassword, newPassword)) {
             System.out.println("Password changed successfully.");
         } else {
             System.out.println("Password change failed. Please check your current password and try again.");
